// src/twitchAPI-types.ts
export interface TwitchUser {
  id: string;
  login: string;
  display_name: string;
  type: string;
  broadcaster_type: string;
  description: string;
  profile_image_url: string;
  offline_image_url: string;
  view_count: number;
  email: string;
  created_at: string;
}
// ...
export interface BotConfig {
  owner_login: string;
  bot_login: string;
  channel_to_moderate: string;
  client_id: string;
  client_secret: string;
}

export interface DBToken {
  access_token: string;
  expires?: number;
  scopes?: string[];
}

export interface TwitchChatters {
  _links: {};
  chatter_count: number;
  chatters: TwitchViewers;
}

export interface TwitchChattersResponse {
  data: TwitchChatters;
}

export interface TwitchViewers {
  broadcaster: string[];
  vips: string[];
  moderators: string[];
  staff: string[];
  admins: string[];
  global_mods: string[];
  viewers: string[];
}

export interface StreamData {
  raiders: string[];
  subs: string[];
  reSubs: string[];
  gifters: any;
  // seedBank: liveData.seedBank ?? [],
  welcomeList: string[];
  cheerers: string[];
  binxRaids: string[];
}